+++ src/common/entity/ItemEntity.hpp
@@ -53,7 +53,7 @@ public:
      * @param y Y坐标
      * @param z Z坐标
      */
-    ItemEntity(EntityId id, const ItemStack& stack, f64 x, f64 y, f64 z);
+    ItemEntity(EntityId id, const ItemStack& stack, f32 x, f32 y, f32 z);
 
     /**
      * @brief 构造物品实体（带投掷速度）
@@ -66,8 +66,8 @@ public:
      * @param vy Y方向速度
      * @param vz Z方向速度
      */
-    ItemEntity(EntityId id, const ItemStack& stack, f64 x, f64 y, f64 z,
-                f64 vx, f64 vy, f64 vz);
+    ItemEntity(EntityId id, const ItemStack& stack, f32 x, f32 y, f32 z,
+                f32 vx, f32 vy, f32 vz);
 
     ~ItemEntity() override = default;
